code cleanup

+++ currencyRates.cpp
@@ -6,15 +6,16 @@ using namespace std;
 class Rates
 {
 public:
-    float usd = 1.0;
-    float euro = 0.94;
-    float peso = 18.36;
-    float gbp = 0.82;
-    float cad = 1.35;
-    float aud = 1.46;
-    float jpy = 134.75;
-    float inr = 82.74;
-    float sgd = 1.33;
-    float chf = 0.92;
-    float cny = 6.88;
+    double usd = 1.0;
+    double euro = 0.94;
+    double peso = 18.36;
+    double gbp = 0.82;
+    double cad = 1.35;
+    double aud = 1.46;
+    double jpy = 134.75;
+    double inr = 82.74;
+    double sgd = 1.33;
+    double chf = 0.92;
+    double cny = 6.88;
+    
 };
